Fix unit tests to check results, and not mock call
