+++ apps/institution-service/cypress/support/utils.ts
@@ -1,5 +1,8 @@
-import { PORT } from "shared/const";
-import { InstitutionAttrs, testInstitution } from "test/testData/institutions";
+import { PORT } from "../../src/shared/const";
+import {
+  InstitutionAttrs,
+  testInstitution,
+} from "../../src/test/testData/institutions";
 import { createAuthorizationHeader } from "../shared/utils/authorization";
 
 interface runTokenInvalidCheckArgs {
@@ -80,26 +83,4 @@ export const createTestInstitution = (token: string) => {
     },
     failOnStatusCode: false,
   });
-};
-
-export const checkIsSorted = (
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  arr: Record<string, any>[],
-  prop: string,
-  direction: "asc" | "desc",
-) => {
-  for (let i = 1; i < arr.length; i++) {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const a = arr[i - 1][prop];
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const b = arr[i][prop];
-
-    if (direction === "asc") {
-      if (a > b) return false;
-    } else {
-      if (a < b) return false;
-    }
-  }
-
-  return true;
 };
